:core 的 Android 读取器各自独立成文件
AppVersion / UserAgent / AcceptLanguage / ImageNetworkPolicyInterceptor 四处的形状是一样的：
中性的值和逻辑，加上边缘一个读设备的函数。拆开不是为了整洁——KMP 里一个文件不能一半
commonMain 一半 androidMain，拆过之后第二阶段就是移动文件而不是改写。

ImageNetworkPolicyInterceptor 顺带去掉了收 Context 的次构造函数，改由 App 传入判断函数；
「什么算未计费网络」本来就是平台的答案（VPN 隧道不报 NOT_METERED 那次排查就是这条）。

:core 里 import android.* 的现在只剩 5 个以平台命名的文件。
